refactor(store): move around exports
