+++ packages/store/package.json
@@ -10,22 +10,26 @@
   "license": "MIT",
   "type": "module",
   "exports": {
-    ".": "./dist/ts/library/index.js",
+    ".": "./dist/ts/index.js",
+    "./codegen": "./dist/ts/codegen/index.js",
+    "./config": "./dist/ts/config/index.js",
     "./register": "./dist/ts/register/index.js",
     "./abi/*": "./abi/*",
-    "./config": "./dist/ts/library/config/index.js",
     "./*": "./dist/*"
   },
   "typesVersions": {
     "*": {
       "index": [
-        "./ts/library/index.ts"
+        "./ts/index.ts"
+      ],
+      "codegen": [
+        "./ts/codegen/index.ts"
+      ],
+      "config": [
+        "./ts/config/index.ts"
       ],
       "register": [
         "./ts/register/index.ts"
-      ],
-      "config": [
-        "./ts/library/config/index.ts"
       ]
     }
   },
@@ -52,6 +56,7 @@
     "@latticexyz/common": "workspace:*",
     "@latticexyz/config": "workspace:*",
     "@latticexyz/schema-type": "workspace:*",
+    "abitype": "0.8.7",
     "ethers": "^5.7.2",
     "zod": "^3.21.4"
   },
