+++ packages/store/ts/config/storeConfig.ts
