+++ packages/store/ts/index.ts
@@ -1,4 +1,6 @@
 // Importing library has no side-effects, unlike register
 // (library neither creates nor extends MUDCoreContext when imported)
 export * from "./config";
-export * from "./render-solidity";
+
+export * from "./storeEvents";
+export * from "./storeEventsAbi";
