+++ packages/store/ts/storeEventsAbi.ts
@@ -1,0 +1,4 @@
+import { parseAbi, AbiEvent } from "abitype";
+import { storeEvents } from "./storeEvents";
+
+export const storeEventsAbi = parseAbi(storeEvents) satisfies readonly AbiEvent[];
