+++ packages/world/ts/plugins/snapsync/configExtensions.ts
@@ -1,5 +1,5 @@
 import { extendMUDCoreConfig, resolveTableId } from "@latticexyz/config";
-import { zPluginStoreConfig } from "@latticexyz/store/config";
+import { zPluginStoreConfig } from "@latticexyz/store/ts/config";
 import { zSnapSyncPluginConfig } from "./plugin";
 import { zPluginWorldConfig } from "../../library";
