+++ packages/world/ts/scripts/tablegen.ts
@@ -1,7 +1,8 @@
 import path from "path";
 import { loadConfig } from "@latticexyz/config/node";
 import { getSrcDirectory } from "@latticexyz/common/foundry";
-import { StoreConfig, tablegen } from "@latticexyz/store";
+import { StoreConfig } from "@latticexyz/store";
+import { tablegen } from "@latticexyz/store/codegen";
 
 const config = (await loadConfig()) as StoreConfig;
 const srcDir = await getSrcDirectory();
